requirements.txt
accelerate==1.10.1
agentops==0.3.26
agentpress==0.1.12
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.20.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
Authlib==1.6.3
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
coloredlogs==15.0.1
comm==0.2.3
cryptography==45.0.6
cyclopts==3.23.1
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docling==2.48.0
docling-core==2.45.0
docling-ibm-models==3.9.0
docling-parse==4.2.3
docstring_parser==0.17.0
docutils==0.22
durationpy==0.10
easyocr==1.7.2
email-validator==2.3.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastmcp==2.11.3
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.28
langchain-core==0.3.74
langchain-docling==1.0.0
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.20
latex2mathml==3.78.0
lazy-object-proxy==1.12.0
lazy_loader==0.4
litellm==1.76.0
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.13.1
mdformat==0.7.22
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.7.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.102.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
ormsgpack==1.10.0
overrides==7.7.0
packaging==23.2
pandas==2.3.2
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-asyncio==0.24.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
questionary==2.1.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
semchunk==2.2.2
setuptools==75.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
streamlit==1.48.1
streamlit-quill==0.0.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==2.4.0
tifffile==2025.6.11
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
